+++ src/action/macos/create_nix_volume.rs
@@ -1,6 +1,6 @@
 use crate::action::{
     base::{create_or_insert_into_file, CreateFile, CreateOrInsertIntoFile},
-    darwin::{
+    macos::{
         BootstrapApfsVolume, CreateApfsVolume, CreateSyntheticObjects, EnableOwnership,
         EncryptApfsVolume, UnmountApfsVolume,
     },
