+++ src/action/macos/create_synthetic_objects.rs
